openclaw-taobao-skill/requirements.txt
playwright>=1.40.0
httpx>=0.25.0
tenacity>=8.2.0
pydantic>=2.0.0
python-dotenv>=1.0.0
pytest>=7.4.0
pytest-asyncio>=0.21.0
